+++ FNPlugin/Science/ComputerCore.cs
@@ -37,7 +37,7 @@ namespace FNPlugin
         [KSPField(isPersistant = true)]
         public bool isupgraded = false;
         [KSPField(isPersistant = true)]
-        public float electrical_power_ratio;
+        public double electrical_power_ratio;
         [KSPField(isPersistant = true)]
         public float last_active_time;
         [KSPField(isPersistant = true)]
@@ -135,12 +135,12 @@ namespace FNPlugin
         {
             if (isupgraded)
             {
-                float power_returned = CheatOptions.InfiniteElectricity 
+                double power_returned = CheatOptions.InfiniteElectricity 
                     ? upgradedMegajouleRate 
                     : consumeFNResource(upgradedMegajouleRate * TimeWarp.fixedDeltaTime, FNResourceManager.FNRESOURCE_MEGAJOULES) / TimeWarp.fixedDeltaTime;
 
                 electrical_power_ratio = power_returned / upgradedMegajouleRate;
-                float altitude_multiplier = (float)(vessel.altitude / (vessel.mainBody.Radius));
+                double altitude_multiplier = vessel.altitude / vessel.mainBody.Radius;
                 altitude_multiplier = Math.Max(altitude_multiplier, 1);
 
                 var scienceMultiplier = PluginHelper.getScienceMultiplier(vessel);
